requirements.txt
streamlit
pandas
reportlab
pillow

